add bill function

+++ backend/controllers/CartController/delete-cart.js
@@ -3,6 +3,15 @@ const deleteCart = async (req, res, next) => {
     try {
         const { id } = req.body;
 
+        console.log(id);
+
+        if (!id) {
+            return res.status(200).json({
+                status: true,
+                message: "Sản phẩm không có trong giỏ!",
+            });
+        }
+
         await db.Cart.destroy({
             where: {
                 id,
